geniusedu/requirements.txt
Package               Version    
--------------------- -----------
alembic               0.9.6      
APScheduler           3.6.3      
asn1crypto            0.24.0     
astroid               2.2.5      
Babel                 2.5.1      
bcrypt                3.1.7      
beautifulsoup4        4.9.1      
blinker               1.4        
boto3                 1.10.15    
botocore              1.13.15    
Brotli                1.0.7      
cachelib              0.1        
certifi               2019.9.11  
cffi                  1.13.2     
chardet               3.0.4      
Click                 7.0        
colorama              0.4.1      
cryptography          2.7        
Django                2.2.5      
docutils              0.15.2     
dominate              2.3.1      
elasticsearch         6.1.1      
eventlet              0.17.4     
Flask                 1.0.2      
Flask-APScheduler     1.11.0     
Flask-Babel           0.11.2     
Flask-Bcrypt          0.7.1      
Flask-Bootstrap       3.3.7.1    
Flask-Caching         1.9.0      
Flask-Compress        1.5.0      
Flask-Cors            3.0.8      
Flask-Dropzone        1.5.4      
Flask-gzip            0.2        
Flask-HTMLmin         1.5.0      
Flask-Login           0.4.0      
Flask-Mail            0.9.1      
Flask-Migrate         2.1.1      
Flask-Minify          0.11       
Flask-Moment          0.5.2      
Flask-Script          2.0.6      
Flask-Session         0.3.2      
Flask-SocketIO        4.2.1      
Flask-SQLAlchemy      2.3.2      
Flask-Static-Compress 1.0.2      
Flask-WTF             0.14.2     
greenlet              0.4.15     
guess-language-spirit 0.5.3      
gunicorn              18.0       
htmlmin               0.1.12     
idna                  2.8        
image                 1.5.27     
isort                 4.3.20     
itsdangerous          0.24       
jac                   0.17.2     
Jinja2                2.10       
jmespath              0.9.4      
jsmin                 2.2.2      
lazy-object-proxy     1.4.1      
lesscpy               0.13.0     
Mako                  1.0.7      
MarkupSafe            1.0        
mccabe                0.6.1      
mysqlclient           1.4.2.post1
ndg-httpsclient       0.5.1      
numpy                 1.18.5     
ordereddict           1.1        
Pillow                6.1.0      
pip                   19.1.1     
ply                   3.11       
pusher                2.1.4      
pyasn1                0.4.7      
pycparser             2.19       
PyJWT                 1.5.3      
pylint                2.3.1      
PyNaCl                1.3.0      
pyOpenSSL             19.0.0     
python-dateutil       2.8.0      
python-dotenv         0.7.1      
python-editor         1.0.3      
python-engineio       3.9.3      
python-slugify        3.0.2      
python-socketio       4.3.1      
pytz                  2017.2     
requests              2.23.0     
rjsmin                1.1.0      
s3transfer            0.2.1      
setuptools            40.8.0     
six                   1.13.0     
slug                  2.0        
soupsieve             2.0.1      
SQLAlchemy            1.1.14     
sqlparse              0.3.0      
stripe                2.48.0     
text-unidecode        1.2        
typed-ast             1.4.0      
tzlocal               2.1        
urllib3               1.25.7     
virtualenv            16.6.1     
visitor               0.1.3      
Werkzeug              0.14.1     
wrapt                 1.11.2     
WTForms               2.1        
xlrd                  1.2.0      
